# docs/partner-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiniLogistics Partner API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5221"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "PartnerApiErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "traceId"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "traceId": { "type": "string" }
            }
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "street": { "type": "string" },
          "ward": { "type": "string" },
          "province": { "type": "string" },
          "country": { "type": "string", "default": "Vietnam" }
        }
      },
      "Parcel": {
        "type": "object",
        "properties": {
          "weightKg": { "type": "number" },
          "lengthCm": { "type": "number" },
          "widthCm": { "type": "number" },
          "heightCm": { "type": "number" }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyBearer": []
    }
  ],
  "paths": {
    "/api/v1/partner/shipping/quote": {
      "post": {
        "summary": "Quote shipping fee",
        "responses": {
          "200": { "description": "Quote calculated" },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PartnerApiErrorResponse" } } } },
          "401": { "description": "Missing or invalid API key" },
          "403": { "description": "Inactive API client" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/v1/partner/shipments": {
      "post": {
        "summary": "Create shipment",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": { "description": "Shipment created" },
          "200": { "description": "Idempotent replay" },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PartnerApiErrorResponse" } } } },
          "401": { "description": "Missing or invalid API key" },
          "403": { "description": "Inactive API client" },
          "409": { "description": "Conflict" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/v1/partner/shipments/{trackingCode}": {
      "get": {
        "summary": "Get shipment tracking",
        "parameters": [
          {
            "name": "trackingCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Shipment found" },
          "401": { "description": "Missing or invalid API key" },
          "403": { "description": "Inactive API client" },
          "404": { "description": "Shipment not found" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/v1/partner/shipments/{trackingCode}/cancel": {
      "post": {
        "summary": "Cancel shipment",
        "parameters": [
          {
            "name": "trackingCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Shipment cancelled" },
          "401": { "description": "Missing or invalid API key" },
          "403": { "description": "Inactive API client" },
          "404": { "description": "Shipment not found" },
          "409": { "description": "Business rule conflict" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    }
  }
}